non-image catalogs failing bug fixed
this bug took a long time to find and a short moment to fix

in the model(), the brightnes is a parameter in linearity,
originally the counts+0.01 was thought to be just an log(0) prevention,
but it turned out to spoil fitting the catalogs that have already
calibrated magnitudes (recalibrations). Fixed by removal and
simplification, should not affect anything.